requirements.txt
selenium>=4.15.0
webdriver-manager>=4.0.0

# PDF Processing and Vector Database
langchain>=0.1.0
langchain-core>=0.1.0
langchain-community>=0.0.20
langchain-experimental>=0.0.50
langchain-openai>=0.0.5
openai>=1.0.0
pinecone>=5.0.0
pypdf>=3.0.0
python-dotenv>=1.0.0
fastapi>=0.109.0
uvicorn[standard]>=0.27.0
pydantic>=2.5.0
pydantic-settings>=2.1.0
python-multipart>=0.0.6
pytesseract>=0.3.10
pillow>=10.2.0